refactor(VZOG-11): 강의 상세 유틸 함수 분리
